move AWS specific code to aws packages
