+++ server/endpoint/creator/aws/aws.go
@@ -1,0 +1,10 @@
+package aws
+
+type Roles struct {
+	Admin       string `json:"admin"`
+	AwsOperator string `json:"awsoperator"`
+}
+
+type AWS struct {
+	Roles Roles `json:"roles"`
+}
