+++ service/creator/request.go
@@ -1,8 +1,11 @@
 package creator
 
+import (
+	"github.com/giantswarm/credentiald/service/creator/aws"
+)
+
 type Request struct {
 	Organization string
 
-	AdminARN       string
-	AwsOperatorARN string
+	AWS aws.Request
 }
